BUG: 4277. Fixing warnings on Visual Studio 8 for Windows 64bits.      Most of them related to conversions between size_t, tsize_t, int,      unsigned int, long and uint32_t types.

+++ Utilities/DICOMParser/DICOMFile.cxx
@@ -84,7 +84,7 @@ void DICOMFile::Close()
 
 long DICOMFile::Tell() 
 {
-  long loc = InputStream.tellg();
+  long loc = static_cast<long>( InputStream.tellg() );
   // dicom_stream::cout << "Tell: " << loc << dicom_stream::endl;
   return loc;
 }
