+++ Utilities/itktiff/tif_jpeg.c
@@ -463,7 +463,7 @@ tables_term_destination(j_compress_ptr cinfo)
   JPEGState* sp = (JPEGState*) cinfo;
 
   /* set tables length to number of bytes actually emitted */
-  sp->jpegtables_length -= sp->dest.free_in_buffer;
+  sp->jpegtables_length -= (uint32)(sp->dest.free_in_buffer);
 }
 
 static int
