+++ Utilities/itktiff/tif_luv.c
@@ -1167,7 +1167,7 @@ LogL16GuessDataFmt(TIFFDirectory *td)
 static uint32
 multiply(size_t m1, size_t m2)
 {
-  uint32  bytes = m1 * m2;
+  uint32  bytes = (uint32)(m1) * (uint32)(m2);
 
   if (m1 && bytes / m1 != m2)
     bytes = 0;
